Factor `repodir()` helper into conftest.py
